+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## master (unreleased)
+* [#202](https://github.com/clojure-emacs/inf-clojure/issues/202): Add ClojureCLR support.
 
 ## 3.2.0 (2022-07-15)
 
@@ -123,4 +124,3 @@
 ### New Features
 
 * [#22](https://github.com/clojure-emacs/inf-clojure/pull/22): Add ElDoc support.
-* [#202](https://github.com/clojure-emacs/inf-clojure/issues/202): Add ClojureCLR support.