+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 * [#202](https://github.com/clojure-emacs/inf-clojure/issues/202): Add ClojureCLR support.
+* [#204](https://github.com/clojure-emacs/inf-clojure/issues/204): Scroll repl buffer on insert commands
+
 
 ## 3.2.1 (2022-07-22)